refac(adm service & handler): adicionar validações nos métodos do Service e try-catch no Handler de Adm

+++ src/app/vanillajavaapi/services/AdmService.java
@@ -29,23 +29,64 @@ public class AdmService {
     }
 
     public AdmResponseDTO findById(String id) {
-        int idFormated = Integer.parseInt(id);
-        Adm adm = admRepository.findById(idFormated);
 
-        return new AdmResponseDTO(adm.getId(), adm.getName(), adm.getEmail());
+        try {
+            int idFormated = Integer.parseInt(id);
+            Adm adm = admRepository.findById(idFormated);
+
+            if(adm == null) {
+                throw new IllegalArgumentException("Adm não encontrado para o ID: " + id);
+            }
+
+            return new AdmResponseDTO(adm.getId(), adm.getName(), adm.getEmail());
+        }
+        catch (NumberFormatException e) {
+            throw new IllegalArgumentException("ID Inválido: " + id);
+        }
     }
 
     public List<AdmResponseDTO> register(String name, String email, String password) {
+
+        if(name == null || name.isBlank()) {
+            throw new IllegalArgumentException("Nome é obrigatório!");
+        }
+
+        if(email == null || email.isBlank() || !email.contains("@")) {
+            throw new IllegalArgumentException("Email faltando ou inválido!");
+        }
+
+        if(password == null || password.isBlank()) {
+            throw new IllegalArgumentException("Senha inválida ou faltando!");
+        }
+
         admRepository.register(name, email, password);
         return this.findAll();
     }
 
     public List<AdmResponseDTO> update(int id, String name, String email, String password) {
-        this.admRepository.update(id, name, email, password);
+
+        Adm existing = admRepository.findById(id);
+
+        if (existing == null) {
+            throw new IllegalArgumentException("Administrador com ID " + id + " não encontrado");
+        }
+
+        if (name != null && !name.isBlank()) existing.setName(name);
+        if (email != null && email.contains("@")) existing.setEmail(email);
+        if (password != null && password.length() >= 6) existing.setPassword(password);
+
+        admRepository.update(id, existing.getName(), existing.getEmail(), existing.getPassword());
         return this.findAll();
     }
 
     public List<AdmResponseDTO> delete(int id) {
+
+        Adm existing = admRepository.findById(id);
+
+        if(existing == null) {
+            throw new IllegalArgumentException("Administrador com ID " + id + " não encontrado!");
+        }
+
         this.admRepository.delete(id);
         return this.findAll();
     }
